requirements.txt
yfinance>=0.2.36
requests>=2.31.0
pandas>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
lxml>=5.0.0
html5lib>=1.1
# Phase 3
psycopg2-binary>=2.9.0
# Phase 5
feedparser>=6.0.0
trafilatura>=1.6.0
vaderSentiment>=3.3.2
# Phase 6
praw>=7.7.0
# Phase 7
anthropic>=0.40.0
fastapi>=0.110.0
uvicorn>=0.27.0
